requirements.txt
streamlit
pytest
psycopg2-binary
python-dotenv
